% Generated by roxygen2 (4.0.2): do not edit by hand
\name{google_dat}
\alias{google_dat}
\title{Get stock values from Google Finance}
\usage{
google_dat(symbols, file_out = NULL)
}
\arguments{
  \item{symbols}{A character vector of symbols to look up
  on Google Finance}

  \item{file_out}{An optional file name to append
  comma-delimited data to. This file must already exist.}
}
\value{
A data frame with current values for each stock symbol. The
first column will contain today's date.
}
\description{
Get stock values from Google Finance
}
\examples{
s <- c("VAB", "XRB", "ZRE", "XEC", "XEF", "XIC",
  "VUN", "TDB909", "TDB911")
google_dat(s)
}

