% Generated by roxygen2 (4.0.2): do not edit by hand
\name{invest}
\alias{invest}
\title{Get investment recommendations}
\usage{
invest(total, current, weights = c(cdn = 0.2, us = 0.2, eur = 0, cdn_b = 0.4,
  int = 0.2), positive_only = FALSE)
}
\arguments{
  \item{total}{Scalar numeric giving the total amount to be
  invested}

  \item{current}{a vector of current assets in the same
  order as \code{weights}}

  \item{weights}{an optionally named vector of desired
  investment weights}

  \item{positive_only}{Should investments be positive only?
  If \code{FALSE} then negative investments may be
  suggested (i.e. selling one asset and purchasing another}

  \item{invest}{the amount to invest}
}
\description{
Get investment recommendations
}
\examples{
invest(30000, c(9000, 7000, 6000, 15000, 7000))
invest(30000, c(9000, 7000, 6000, 15000, 7000), positive_only = TRUE)
}

