% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mdietz}
\alias{mdietz}
\title{Modified Dietz method}
\usage{
mdietz(start_value, end_value, f, f_day, days)
}
\arguments{
  \item{start_value}{The starting value}

  \item{end_value}{The ending value}

  \item{f}{A vector of in and out cash flows}

  \item{f_day}{A numeric vector of days since start in
  which cash flows occured}

  \item{days}{The (numeric) total number of days (end day
  minus first day + 1)}
}
\value{
A percent return
}
\description{
Modified Dietz method
}
\examples{
mdietz(start_value = 100, end_value = 300, f = c(50),
   f_day = c(365), days <- 365*2)

 mdietz(start_value = 100, end_value = 300, f = c(50, 150),
   f_day = c(365, 380), days <- 365*2)

 mdietz(start_value = 100, end_value = 300, f = c(50, 10),
   f_day = c(365, 380), days <- 365*2)
}
\references{
http://en.wikipedia.org/wiki/Modified_Dietz_method
}

