% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ror}
\alias{ror}
\title{Get rate of return}
\usage{
ror(investment_net, contributions, buys)
}
\arguments{
  \item{investment_net}{Net value invested}

  \item{contributions}{A vector of contributions}

  \item{buys}{A vector of value bought}
}
\value{
A scalar numeric value representing the rate of return
}
\description{
Get rate of return
}

