% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rss}
\alias{rss}
\title{Calculate residual sums of squares for weights and desired weights}
\usage{
rss(inv_weights, desired_dist, total, current)
}
\arguments{
  \item{inv_weights}{a vector of weights. Should be missing
  the first weight.}

  \item{desired_dist}{the desired weights. Should be one
  longer than inv_weights.}

  \item{total}{total amount to invest}

  \item{current}{amounts invested in the assets}
}
\description{
Calculate residual sums of squares for weights and desired
weights
}

